+++ example/src/demo/VirtualizedSelect.example.js
@@ -4,19 +4,26 @@ import styles from './VirtualizedSelect.module.css';
 import {
   ContentBox,
 } from '../common/ContentBox';
+import { FormLabel } from '../common/Form';
 
-const cityData = [];
+const getRows = (rowCount = 0) => {
+  const rows = [];
 
-for (let i = 0; i < 100000; i += 1) {
-  cityData.push({
-    value: `item-${i}`,
-    label: `Item ${i}`,
-  });
+  for (let i = 0; i < rowCount; i += 1) {
+    rows.push({
+      value: `item-${i}`,
+      label: `Item ${i}`,
+    });
+  }
+
+  return rows;
 }
 
 export default class VirtualizedSelectExample extends React.Component {
   constructor(props) {
     super(props);
+
+    const defaultRowCount = 1000;
 
     this.state = {
       clearable: true,
@@ -24,21 +31,37 @@ export default class VirtualizedSelectExample extends React.Component {
       multi: false,
       searchable: true,
       selectedCity: '',
+      rowCount: defaultRowCount,
+      rows: getRows(defaultRowCount)
     };
   }
 
   render() {
-    const { clearable, disabled, multi, searchable, selectedCity } = this.state;
+    const { clearable, disabled, multi, searchable, selectedCity, rows, rowCount } = this.state;
 
     return (
       <ContentBox>
+        <FormLabel label="rowCount" inline style={{ margin: '5px 0 16px 0' }}>
+          <input
+            name="rowCount"
+            value={rowCount}
+            onChange={event => {
+              const newRowCount = Number(event.target.value);
+              this.setState({
+                rowCount: newRowCount,
+                rows: getRows(newRowCount),
+              });
+            }}
+          />
+        </FormLabel>
+
         <div style={{ width: 500 }}>
           <VirtualizedSelect
             isClearable={clearable}
             isDisabled={disabled}
             isSearchable={searchable}
             onChange={selectedCity => this.setState({ selectedCity })}
-            options={cityData}
+            options={rows}
             value={selectedCity}
             isMulti={multi}
             valueKey="value"